Implement Nokia's DirectGraphics and add methods to read from  screen
SquirrelJME now supports Nokia DirectGraphics and screen reading!

DirectGraphics wasn't implemented yet, and is used by quite a few
J2ME applications, even those not actually meant to run on Nokia
devices, like some Sony Ericsson ones.

This commit solves that and also adds some new methods for reading
and drawing to the screen from any supported pixel format:
- drawRegion()
- getRegion()
Both of which are now available to the Java side.

Also, it is now possible to access the native pixel format of the
pencil from Java, as it is required by DirectGraphics.

Some reworks were done to SquirrelJME's base Graphics stack as well,
such as the addition of an 'ExtraGraphics' interface, to be
extended with more vendor-specific operations in the future.